requirements.txt
aiogram==3.11.0
APScheduler==3.10.4
beautifulsoup4==4.12.3
django_environ==0.11.2
Faker==26.0.0
loguru==0.7.2
python-decouple==3.8
Requests==2.32.3
SQLAlchemy==2.0.31
redis==5.0.8
cachetools==5.4.0